requirements.txt
alembic==1.4.2
Authlib==0.14.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cryptography==3.1
ecdsa==0.14.1
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jose==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
